renamed ResourceSystem to resource_system because no capital letter in package names!!!
