+++ core/src/main/java/com/strategy_bit/chaos_brawl/player_input_output/PawnController.java
@@ -1,8 +1,8 @@
 package com.strategy_bit.chaos_brawl.player_input_output;
 
 
-import com.strategy_bit.chaos_brawl.ResourceSystem.Resource;
-import com.strategy_bit.chaos_brawl.ResourceSystem.ResourceGold;
+import com.strategy_bit.chaos_brawl.resource_system.Resource;
+import com.strategy_bit.chaos_brawl.resource_system.ResourceGold;
 import com.strategy_bit.chaos_brawl.types.UnitType;
 import com.strategy_bit.chaos_brawl.util.Boundary;
 import com.strategy_bit.chaos_brawl.world.InputHandler;
