+++ core/src/main/java/com/strategy_bit/chaos_brawl/player_input_output/PlayerController.java
@@ -3,7 +3,7 @@ package com.strategy_bit.chaos_brawl.player_input_output;
 import com.badlogic.gdx.InputProcessor;
 import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.scenes.scene2d.Stage;
-import com.strategy_bit.chaos_brawl.ResourceSystem.Resource;
+import com.strategy_bit.chaos_brawl.resource_system.Resource;
 import com.strategy_bit.chaos_brawl.types.UnitType;
 import com.strategy_bit.chaos_brawl.util.Boundary;
 import com.strategy_bit.chaos_brawl.player_input_output.views.GameHUD;
