+++ core/src/main/java/com/strategy_bit/chaos_brawl/world/Board.java
@@ -9,7 +9,7 @@ import com.badlogic.gdx.math.Vector2;
 public interface Board {
     //public BoardType getBoardType();
     public Tile[][] getTileBoard();
-
+    //TODO Florian let the board
     /**
      * returns a matrix like<br>
      * current implementation 0 entries stays for not walkable and 1 for walkable<br>
